Added Bootstrap and Added to Create Page
Created the create.html page from scratch again using bootstrap. Also updated the navbar to use bootstrap. Navbar currently needs a bit more work since the navbar collapse is not working properly after my changes, but that is only an issue for mobile responsiveness. Changed some linking of files as well so hopefully that is working.