Styling fixes and adding browse button to homepage
I went through the login and create account pages and reset password page and fixed up the styling some more to match eachother and the rest of the website. I also added an option to browse trips from the main homepage below the big image. For each of the pages I also did some mobile responsiveness fixes so that should be working fairly well. There are still some fixes to be added in terms of responsiveness but overall pages seem good. (apart from the profile page which still needs styles)