Fixed up issues with certain route cases
I went through and changed up a fair amount of the code, seeing as it was leading to a lot of errors with certain types of trips. Adding hotels/gas/food should be working now but I still need to make it so that these add as waypoints on the route. This will require a bit of a workaround with how the routes and waypoints work